Adding CORS
In this commit, a CORS middleware is established to manage Cross-Origin requests within the candidate careers web application. Within this middleware, the Access-Control-Allow-Origin header is configured as '*' to enable future access to this API from any URL or domain.

Through the inclusion of this header within the middleware, the candidate careers web application gains the ability to interact with resources located on distinct domains while upholding necessary security protocols.